requirements.txt
alpaca-py>=0.33.0
streamlit>=1.39.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
sqlalchemy>=2.0.0
scipy>=1.11.0
numpy>=1.26.0
libsql-experimental>=0.0.40
